+++ CHANGES.rst
@@ -2,6 +2,15 @@
 Change History
 ==============
 
+
+0.2.1 (2013-12-17)
+------------------
+
+- Added tests.
+- Fixed comment gitignore patterns.
+- Fixed relative path gitignore patterns.
+
+
 0.2.0 (2013-12-07)
 ------------------